pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "LogLead"
version = "1.0.0-beta"
authors = [
    {name = "Mika Mäntylä", email="mika.mantyla@helsinki.fi"},
    {name = "Jesse Nyyssölä", email = "jesse.nyyssola@helsinki.fi"},
    {name = "Yuqing Wang", email="yuqing.wang@helsinki.fi"},
    {name = "Alexander Bakhtin", email = "alexander.bakhtin@protonmail.com"}]
maintainers = [
    {name = "Alexander Bakhtin", email = "alexander.bakhtin@protonmail.com"}]
description = "LogLead stands for Log Loader, Enhancer, and Anomaly Detector"
readme = "PYPI_README.md"
license = {file = "LICENSE"}
keywords = [
    "logs",
    "anomaly detection",
    "log parsing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging"]

requires-python = ">=3.9"
dependencies = [
"polars[numpy, pandas, pyarrow]>=1.5",
"scipy>=1.10.1",
'regex>=2023.10.3',
'drain3>=0.9.11',
'tipping>=0.1.3',
"scikit-learn>=1.2.2",
"xgboost>=1.7.3",
"python-dotenv>=1.0.1",
"pyyaml",
"psutil",
"jinja2",
"matplotlib",
"requests", #tests 
"tqdm", #tests
"GitPython", #tests
"py7zr", #tests 
"shap>=0.42.1", #explainer 
"umap-learn>=0.5.6", #explainer
"plotly>=5.19.0", #explainer
"nbformat>=4.2.0", #explainer
"xlsxwriter" #Robo mode 
]

[project.urls]
Homepage = "https://github.com/EvoTestOps/LogLead"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["loglead*"]
